requirements.txt
appdirs==1.4.4
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
blessed==1.19.1
bpython==0.23
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.1.2
cryptography==39.0.0
curtsies==0.4.1
cwcwidth==0.1.8
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
frozendict==2.3.4
greenlet==2.0.1
html5lib==1.1
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
lxml==4.9.2
matplotlib-inline==0.1.6
multitasking==0.0.11
mypy==0.991
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.230105
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.8.0
types-pytz==2022.7.0.0
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
yfinance==0.2.3
